requirements.txt
numpy<2
pypdf
python-dotenv
langchain
langchain-community
langchain-pinecone
langchain-huggingface
langchain-text-splitters
sentence-transformers==2.7.0
transformers==4.44.0
pinecone
huggingface_hub
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu